Marking: type the local entry count as int32_t, not size_t
Match the repo convention (e.g. fem/interpolate.h, mesh/utils.h) of
storing index_map.size_local() + index_map.num_ghosts() as int32_t,
and only widen to size_t at the point of comparison against
values.size().

+++ cpp/dolfinx/refinement/mark.h
@@ -69,9 +69,8 @@ std::vector<std::int32_t> mark_maximum(std::span<const T> values,
         std::format("theta must satisfy 0 < theta <= 1, got {}.", theta));
   }
 
-  const std::size_t size = static_cast<std::size_t>(index_map.size_local()
-                                                    + index_map.num_ghosts());
-  if (values.size() != size)
+  const std::int32_t size = index_map.size_local() + index_map.num_ghosts();
+  if (values.size() != static_cast<std::size_t>(size))
   {
     throw std::invalid_argument(
         std::format("values must have size index_map.size_local() + "
